refactor(nika-display): the models rung moves out at the 1500-line wall
The judged-defaults merge grew check_render.rs to 1,567 lines — over
the file cap (hygiene vector 12). The MODELS section (ModelFinding ·
ModelsAudit · the row builder) moves to a private check_models module;
the re-export keeps the public path check_render::{ModelFinding,
ModelsAudit} byte-stable (public-api rendering grep-verified: zero
check_models paths leak). 1,435 lines remain.
