+++ crates/nika-display/src/lib.rs
@@ -19,6 +19,7 @@
 // nika-cli crate holds at its root — inherited by the descent).
 #![cfg_attr(test, allow(clippy::unwrap_used, clippy::expect_used))]
 
+mod check_models;
 pub mod check_render;
 pub mod chrome;
 mod claims;
